Got basic input port working.

+++ Source/Peripheral_Libs/ICs.MCP23008/Driver/ICs.MCP23008/DigitalOutputPort.cs
@@ -9,7 +9,7 @@ namespace Netduino.Foundation.ICs.MCP23008
     /// </summary>
     public class DigitalOutputPort : DigitalOutputPortBase
     {
-        protected readonly int _pin;
+        protected readonly byte _pin;
         protected readonly MCP23008 _mcp;
 
         public override bool State
@@ -28,7 +28,7 @@ namespace Netduino.Foundation.ICs.MCP23008
 
         protected DigitalOutputPort() : base(false) { }
 
-        internal DigitalOutputPort(MCP23008 mcp, int pin, bool initialState) : base(initialState)
+        internal DigitalOutputPort(MCP23008 mcp, byte pin, bool initialState) : base(initialState)
         {
             _mcp = mcp;
             _pin = pin;
